API complété
40% - Multer et Authorize

70% - Model Sauce, Route Sauce, Controller Sauce

Ajouter/Supprimer une sauce et Voir toutes les sauces

100% - Modif Sauce avec ou sans image, Like/Dislike ou neutre, Modif ou Suppression par Propriétaire seulement

+++ controllers/sauce.js
@@ -52,32 +52,41 @@ exports.createSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
-    //On récupère l'objet voulu
-    Sauce.findOne({ _id: req.params.id })
-        .then(sauce => {
-            //On fragmente l'url pour extraire le nom du fichier
-            const filename = sauce.imageUrl.split('/images/')[1];
-            //fs.unlink supprime le fichier et envoi un callback a executé ensuite (Sauce.deleteOne)
-            fs.unlink(`images/${filename}`, () => {
-                //Création de l'objet sauceObject qui vérifie si req.file existe (avec un raccourci de if {} else {})
-                const sauceObject = req.file 
-                    //S'il existe, on traite la nouvelle image
-                    ? {
+    //Si un fichier est transmis
+    if (req.file) {
+        //On récupère l'objet voulu
+        Sauce.findOne({ _id: req.params.id })
+            .then(sauce => {
+                //On fragmente l'url pour extraire le nom du fichier
+                const filename = sauce.imageUrl.split('/images/')[1];
+                //fs.unlink supprime le fichier et envoi un callback a executé ensuite (Sauce.updateOne)
+                fs.unlink(`images/${filename}`, () => {
+                    //Création de l'objet sauceObject après vérification de l'existence de req.file
+                    const sauceObject = {
                         ...JSON.parse(req.body.sauce),
                         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-                    } 
-                    //Sinon, on traite seulement l'objet entrant
-                    : {
-                        ...req.body
                     };
-                
-                //Récupère le produit par son /:id et le remplace par le nouveau en vérifiant qu'il s'agisse bien de ce même id
-                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-                    .then(() => res.status(200).json({message: "Objet modifié !"}))
-                    .catch(error => res.status(400).json({ error }));
+                    
+                    // Récupère le produit par son /:id et le remplace par le nouveau en vérifiant qu'il s'agisse bien de ce même id
+                    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                        .then(() => res.status(200).json({message: "Objet modifié !"}))
+                        .catch(error => res.status(400).json({ error }));
                 })
-        })
-        .catch( error => res.status(500).json({ error }));
+            })
+            .catch( error => res.status(500).json({ error }));
+    }
+
+    else {
+        //Création de l'objet sauceObject après vérification de l'existence de req.file
+        const sauceObject = {
+            ...req.body
+        };
+        
+        // Récupère le produit par son /:id et le remplace par le nouveau en vérifiant qu'il s'agisse bien de ce même id
+        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+            .then(() => res.status(200).json({message: "Objet modifié !"}))
+            .catch(error => res.status(400).json({ error }));
+    }
 };
 
 exports.deleteSauce = (req, res, next) => {
@@ -147,3 +156,57 @@ exports.getAllSauces = (req, res, next) => {
 //     ];
 //     res.status(200).json(sauce);
 // };
+
+exports.likeSauce = (req, res, next) => {
+    console.log("------------------------------");
+
+    console.log("req.body :", req.body);
+
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+            const sauceObject = {};
+
+            console.log("sauce BEFORE", sauce);
+
+            //Si on like
+            if (req.body.like == 1) {
+                //On ajoute (en 2 temps, sinon = .length) l'id de l'utilisateur qui like
+                sauce.usersLiked.push(req.body.userId);
+
+                //On attribut à l'objet les nouvelles infos
+                sauceObject.usersLiked = sauce.usersLiked;
+                sauceObject.likes = sauce.usersLiked.length;
+            }
+
+            //Sinon, si on retire le like
+            else if (req.body.like == 0) {
+                //On enlève (en 2 temps) l'id des Likes ou des Dislikes
+                const newUsersLiked = sauce.usersLiked.filter(id => id !== req.body.userId);
+                const newUsersDisliked = sauce.usersDisliked.filter(id => id !== req.body.userId);
+
+                //On attribut à l'objet les nouvelles infos
+                sauceObject.usersLiked = newUsersLiked;
+                sauceObject.usersDisliked = newUsersDisliked;
+                sauceObject.likes = newUsersLiked.length;
+                sauceObject.dislikes = newUsersDisliked.length;
+            }
+
+            //Sinon, si on dislike
+            else if (req.body.like == -1) {
+                //On ajoute (en 2 temps, sinon = .length) l'id de l'utilisateur qui dislike
+                sauce.usersDisliked.push(req.body.userId);
+
+                //On attribut à l'objet les nouvelles infos
+                sauceObject.usersDisliked = sauce.usersDisliked;
+                sauceObject.dislikes = sauce.usersDisliked.length;
+            }
+
+            console.log("sauceObject :", sauceObject);
+
+            // Récupère le produit par son /:id et le remplace par le nouveau en vérifiant qu'il s'agisse bien de ce même id
+            Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                .then(() => res.status(200).json({message: "Objet modifié !"}))
+                .catch(error => res.status(400).json({ error }));
+        })
+        .catch( error => res.status(500).json({ error }));
+};
